+++ jit/internal/kind.go
@@ -69,7 +69,7 @@ func init() {
 	SetCpuWidth(CpuWidthHost)
 }
 
-type Size uintptr // 1, 2, 4, 8 or 16
+type Size uint64 // 1, 2, 4, 8 or 16
 
 // intentionally matches reflect.Kind values
 type Kind uint8
